bpf: add support for wildcard map
Joe and I are working on the wildcard map support and below is a quick
way of reusing the existing bpf hash table to achieve the wildcard map.
The idea is for user API to have the map's key as {ukey, mask}, see
struct wckey, and we save the key={ukey&mask, mask} into the bpf hash
table.

When lookup, ukey will be masked by each existing table entry's mask
and compare against the masked key.  If matched, then we have a hit.
For update and delete, apply the mask to the ukey then passes to
htab_map_{update, delete}_elem(). See test_maps.c for example.
For lookup, since we have to iterate through all element in the hash
table, the complxity is O(n) thus performance isn't so good.

hash_lookup 55,026,341
wc_lookup    8,892,035

Joe is working on improving the underlying data structure, as well
as refactoring the OVS's wildcard table, called tuple space. Once
finished and evaluated, we can consider switching to use tuple space.

Thanks!

./map_perf_test
  0:hash_map_perf pre-alloc 3715245 events per sec
  0:percpu_hash_map_perf pre-alloc 3531099 events per sec
  0:hash_map_perf kmalloc 2743004 events per sec
  0:percpu_hash_map_perf kmalloc 185164 events per sec
  0:lru_hash_map_perf pre-alloc 2614592 events per sec
  0:nocommon_lru_hash_map_perf pre-alloc 2882798 events per sec
  0:lpm_perf kmalloc 2870770 events per sec
  0:hash_lookup 55026341 lookups per sec
  0:array_lookup 124435007 lookups per sec
  0:inner_lru_hash_map_perf pre-alloc 2784668 events per sec

[wildcard map]
  0:wcmap_perf pre-alloc 1313470 events per sec
  0:wcmap_lookup 8892035 lookups per sec